[completes #53397511] User sees pretty time in task list
I moved pretty_time helper to the application_helper and organized the
methods alphabetically.  I also cleaned up some whitespace in the
notifications spec.
